Move to organization level julia ci/cd

+++ src/acsets/DenseACSets.jl
@@ -212,7 +212,7 @@ struct AnonACSet{S,Ts,Parts,Subparts} <: StructACSet{S,Ts}
 
   function AnonACSet{S,Ts,Parts,Subparts}() where {S,Ts,Parts,Subparts}
     new{S,Ts,Parts,Subparts}(
-      Parts(zeros(Int64, length(S.parameters[2].parameters))),
+      Parts(zeros(Int, length(S.parameters[2].parameters))),
       Subparts(T() for T in Subparts.parameters[2].parameters)
     )
   end
